one more MacOSX compilation fix after commit #9721
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@9726 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/src/ParametricCurves.cpp
@@ -239,7 +239,7 @@ static void computeCatmullRomGlobalParameter(const vector<Coord> &controlPoints,
   float totalDist = 0.f;
 
   for (size_t i = 1 ; i < controlPoints.size() ; ++i) {
-    float dist = pow(controlPoints[i-1].dist(controlPoints[i]), alpha);
+    float dist = pow((float) controlPoints[i-1].dist(controlPoints[i]), alpha);
     cumDist[i] = cumDist[i-1] + dist;
     totalDist += dist;
   }
